move db models to /db
