+++ src/db/index.ts
@@ -1,5 +1,7 @@
 export { connectDb } from './connect';
+
 export { Portfolio, IPortfolioDocument } from './portfolio/Portfolio';
+
 export { 
   StockTransaction, 
   OptionTransaction, 
@@ -9,3 +11,13 @@ export {
   IStockTransaction, 
   IStockTransactionDocument 
 } from './transaction';
+
+export {
+  AccountActivityItem,
+  AccountActivityItemDocument
+} from './robinhood/AccountActivityItem';
+
+export {
+  PortfolioSummaryItem,
+  PortfolioSummaryItemDocument
+} from './robinhood/PortfolioSummaryItem';
